mt6735: Don't allocate hwc displays for virtual displays
Newer versions of the kernel and bootloader will render displays using
reverse scan to mitigate possible screen flickering, this means that the
framebuffer is rendered upside down. To compensate for this, the display
subsystem flips it back to a natural position. However, virtual displays
that use hardware composer will still be upside down.

This change makes virtual displays renderable using the gpu and GLES
instead of going through the hardware composer.

Change-Id: I8ffb2eab7c862783f34ac97c214eb03467e587fa
Ticket: PORRIDGE-492